Move nested query containers into model directory
